+++ app/Main.hs
@@ -10,11 +10,13 @@ import Console.Parse (parseTimestamps)
 import Console.Report (ReportData (..), writeReport)
 import Console.Types (diffElapsed, getInterval, getLogLine, tciTimeElapsed)
 import qualified Data.Text.IO as T
+import Data.Time.Clock (getCurrentTime)
 
 main :: IO ()
 main = do
     log_ts <- parseTimestamps =<< T.readFile "consoleText_timestamps"
     testInfos <- readMethodDurations "junitreport"
+    generatedOn <- getCurrentTime
     let log_plain = getLogLine <$> log_ts
         buildDuration = uncurry diffElapsed $ getInterval log_ts
         slowTestMethods = takeWhile (\methodInfo -> mdDuration methodInfo > 20) testInfos
@@ -27,4 +29,5 @@ main = do
             mavenData
             buildDuration
             pluginStats
+            generatedOn
     writeReport reportData "report.html"