+++ app/Main.hs
@@ -4,11 +4,12 @@ module Main where
 
 import Console.Checks.MavenDownload (getMavenDownloadData)
 import Console.Checks.MavenPlugin (getPluginStats)
+import Console.Checks.RepoDuration (getBuildDurationPerRepo)
 import Console.Checks.TestDuration (getClassInfos, mdDuration,
                                     readMethodDurations)
 import Console.Parse (parseTimestamps)
 import Console.Report (ReportData (..), writeReport)
-import Console.Types (diffElapsed, getInterval, getLogLine, tciTimeElapsed)
+import Console.Types (getDuration, getLogLine, tciTimeElapsed)
 import qualified Data.Text.IO as T
 import Data.Time.Clock (getCurrentTime)
 
@@ -18,16 +19,18 @@ main = do
     testInfos <- readMethodDurations "junitreport"
     generatedOn <- getCurrentTime
     let log_plain = getLogLine <$> log_ts
-        buildDuration = uncurry diffElapsed $ getInterval log_ts
-        slowTestMethods = takeWhile (\methodInfo -> mdDuration methodInfo > 20) testInfos
         slowTestClasses = takeWhile (\classInfo -> tciTimeElapsed classInfo > 60) $ getClassInfos log_plain
-        pluginStats = getPluginStats log_ts
+        slowTestMethods = takeWhile (\methodInfo -> mdDuration methodInfo > 20) testInfos
         mavenData = getMavenDownloadData log_ts
+        buildDuration = getDuration log_ts
+        pluginStats = getPluginStats log_ts
+        durationPerRepo = getBuildDurationPerRepo log_ts
         reportData = ReportData
             slowTestClasses
             slowTestMethods
             mavenData
             buildDuration
             pluginStats
+            durationPerRepo
             generatedOn
     writeReport reportData "report.html"